refactor: things spotted in pr review
